FEATURE Reorganizing folders for MainActivity and MainApplication
